requirements.txt
numpy
fastapi
uvicorn
opencv-python
requests
torch
torchvision
faiss-cpu
pydantic-settings